timestamps to uint32 for packing

+++ contracts/interfaces/IDeployer.sol
@@ -24,7 +24,7 @@ struct DeploymentParams {
     RevenueShare dist; // revenue sharing splits between RToken and RSR
     //
     // === Rewards (Furnace + StRSR) ===
-    uint256 rewardPeriod; // {s} the atomic unit of rewards, determines # of exponential rounds
+    uint32 rewardPeriod; // {s} the atomic unit of rewards, determines # of exponential rounds
     int192 rewardRatio; // the fraction of available revenues that stRSR holders get each PayPeriod
     //
     // === StRSR ===
